+++ src/components/Grants/GrantsPastItem.tsx
@@ -5,18 +5,19 @@ import { navigate } from 'decentraland-gatsby/dist/plugins/intl'
 import Time from 'decentraland-gatsby/dist/utils/date/Time'
 import { Table } from 'decentraland-ui/dist/components/Table/Table'
 
-import { GrantAttributes, ProposalGrantCategory, ProposalGrantCategoryColor } from '../../entities/Proposal/types'
+import { GrantAttributes, ProposalGrantCategory } from '../../entities/Proposal/types'
 import locations from '../../modules/locations'
 import Pill from '../Common/Pill'
 import ChevronRight from '../Icon/ChevronRight'
 
+import { PROPOSAL_GRANT_CATEGORY_COLORS } from './GrantCard'
 import './GrantsPastItem.css'
 
 const GrantsPastItem = ({ grant }: { grant: GrantAttributes }) => {
   const t = useFormatMessage()
   const { id, title, configuration, start_at } = grant
   const category = configuration.category.split(' ')[0]
-  const pillColor = ProposalGrantCategoryColor[configuration.category as ProposalGrantCategory]
+  const pillColor = PROPOSAL_GRANT_CATEGORY_COLORS[configuration.category as ProposalGrantCategory]
   const handleClick = useCallback(() => navigate(locations.proposal(id)), [id])
 
   return (
